added AuthorityService permissions logic in ExerciseService #56

+++ src/main/java/com/behabits/gymbo/application/service/ExerciseServiceImpl.java
@@ -3,6 +3,8 @@ package com.behabits.gymbo.application.service;
 import com.behabits.gymbo.domain.daos.ExerciseDao;
 import com.behabits.gymbo.domain.models.Exercise;
 import com.behabits.gymbo.domain.models.Serie;
+import com.behabits.gymbo.domain.models.User;
+import com.behabits.gymbo.domain.services.AuthorityService;
 import com.behabits.gymbo.domain.services.ExerciseService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,34 +16,46 @@ import java.util.List;
 public class ExerciseServiceImpl implements ExerciseService {
 
     private final ExerciseDao exerciseDao;
+    private final AuthorityService authorityService;
 
     @Override
     public Exercise findExerciseById(Long id) {
-        return this.exerciseDao.findExerciseById(id);
+        Exercise exercise = this.exerciseDao.findExerciseById(id);
+        this.authorityService.checkLoggedUserHasPermissions(exercise);
+        return exercise;
     }
 
     @Override
     public Exercise createExercise(Exercise exercise) {
+        User user = this.authorityService.getLoggedUser();
+        exercise.setUser(user);
         return this.exerciseDao.createExercise(exercise);
     }
 
     @Override
     public List<Exercise> findExercisesByTrainingId(Long trainingId) {
-        return this.exerciseDao.findExercisesByTrainingId(trainingId);
+        User user = this.authorityService.getLoggedUser();
+        return this.exerciseDao.findExercisesByTrainingIdAndUserId(trainingId, user.getId());
     }
 
     @Override
     public List<Serie> findSeriesByExerciseId(Long exerciseId) {
+        Exercise exercise = this.findExerciseById(exerciseId);
+        this.authorityService.checkLoggedUserHasPermissions(exercise);
         return this.exerciseDao.findSeriesByExerciseId(exerciseId);
     }
 
     @Override
     public Serie createSerie(Long exerciseId, Serie serie) {
+        Exercise exercise = this.findExerciseById(exerciseId);
+        this.authorityService.checkLoggedUserHasPermissions(exercise);
         return this.exerciseDao.createSerie(exerciseId, serie);
     }
 
     @Override
     public void deleteExercise(Long id) {
+        Exercise exercise = this.findExerciseById(id);
+        this.authorityService.checkLoggedUserHasPermissions(exercise);
         this.exerciseDao.deleteExercise(id);
     }
 }
